Revert "Feature/home controller"
