+++ .gitignore
@@ -10,5 +10,3 @@ Homestead.yaml
 npm-debug.log
 yarn-error.log
 .env
-
-public/report
