+++ _ide_helper.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * A helper file for Laravel 5, to provide autocomplete information to your IDE
- * Generated for Laravel 5.6.24 on 2018-09-01 17:22:26.
+ * Generated for Laravel 5.6.24 on 2018-09-01 07:42:01.
  *
  * This file should not be included in your code, only analyzed by your IDE!
  *
