+++ app/Game.php
@@ -6,10 +6,6 @@ use Illuminate\Database\Eloquent\Model;
 
 class Game extends Model
 {
-    protected $casts = [
-        'user_id' => 'int'
-    ];
-
     protected $fillable = [
         'complete',
         'score',
@@ -25,18 +21,13 @@ class Game extends Model
         });
     }
 
-    public function frames()
-    {
-        return $this->hasMany(Frame::class);
-    }
-
     public function path()
     {
         return '/games/' . $this->id;
     }
 
-    public function rolls()
+    public function frames()
     {
-        return $this->hasManyThrough(Roll::class, Frame::class);
+        return $this->hasMany(Frame::class);
     }
 }
