+++ app/Http/Controllers/Auth/ForgotPasswordController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\SendsPasswordResetEmails;
 
-/** @codeCoverageIgnore */
 class ForgotPasswordController extends Controller
 {
     /*
