+++ app/Http/Controllers/Auth/LoginController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
 
-/** @codeCoverageIgnore */
 class LoginController extends Controller
 {
     /*
