+++ app/Http/Controllers/Auth/RegisterController.php
@@ -8,7 +8,6 @@ use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Foundation\Auth\RegistersUsers;
 
-/** @codeCoverageIgnore */
 class RegisterController extends Controller
 {
     /*
