+++ app/Http/Controllers/Auth/ResetPasswordController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\ResetsPasswords;
 
-/** @codeCoverageIgnore */
 class ResetPasswordController extends Controller
 {
     /*
