+++ app/Roll.php
@@ -6,7 +6,5 @@ use Illuminate\Database\Eloquent\Model;
 
 class Roll extends Model
 {
-    protected $fillable = [
-        'pins'
-    ];
+    //
 }
