+++ app/User.php
@@ -32,14 +32,4 @@ class User extends Authenticatable
     {
         return $this->hasMany(Game::class);
     }
-
-    public function getFullNameAttribute()
-    {
-        return "{$this->first_name} {$this->last_name}";
-    }
-
-    public function path()
-    {
-        return "/users/{$this->id}";
-    }
 }
