--- resources/views/games/create.blade.php
@@ -1,3 +1,0 @@
-<h2>Create Game</h2>
-<form action="{{ route('games.store') }}" id="game-create-form" method="post">
-</form>
