--- resources/views/games/edit.blade.php
@@ -1,1 +1,0 @@
-Edit Game {{ $game->id }}
