--- resources/views/users/index.blade.php
@@ -1,3 +1,0 @@
-@foreach($users as $user)
-    {{ $user->fullName }}
-@endforeach
