--- resources/views/users/show.blade.php
@@ -1,1 +1,0 @@
-{{ $user->fullName }}'s Profile
