+++ tests/Unit/GameTest.php
@@ -4,7 +4,6 @@ namespace Tests\Unit;
 
 use App\Frame;
 use App\Game;
-use App\Roll;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Tests\TestCase;
 
@@ -33,35 +32,6 @@ class GameTest extends TestCase
     }
 
     /** @test */
-    public function it_should_show_authenticated_users_the_game_creation_form()
-    {
-        $this->signIn();
-
-        $response = $this->get('/games/create');
-
-        $response->assertSee('<form');
-        $response->assertSee('id="game-create-form"');
-    }
-
-    /** @test */
-    public function guests_may_not_see_the_game_create_form()
-    {
-        $this->expectException(\Illuminate\Auth\AuthenticationException::class);
-
-        $this->get(route('games.create'))
-             ->assertRedirect('/login');
-    }
-
-    /** @test */
-    public function it_should_show_authenticated_users_the_edit_form_for_a_game()
-    {
-        $this->signIn();
-        $game = create(Game::class);
-
-        $this->get(route('games.edit', $game->id))->assertSee('Edit Game ' . $game->id);
-    }
-
-    /** @test */
     public function it_should_show_a_single_game(): void
     {
         $game = create(Game::class);
@@ -71,11 +41,35 @@ class GameTest extends TestCase
     }
 
     /** @test */
+    public function it_should_store_a_game()
+    {
+        $this->signIn();
+        $game = make(Game::class);
+        unset($game->user_id);
+
+        $this->post($game->path(), $game->toArray());
+
+        $this->assertDatabaseHas('games', $game->toArray());
+    }
+
+    /** @test */
     public function a_game_should_have_a_score()
     {
         $game = create(Game::class);
 
         $this->assertArrayHasKey('score', $game->toArray());
+    }
+
+    /** @test */
+    public function stored_games_should_belong_to_the_currently_authenticated_user()
+    {
+        $this->signIn();
+        $game = make(Game::class);
+
+        $this->post('games', $game->toArray());
+        $usersGame = $this->user->games()->first();
+
+        $this->assertEquals($this->user->id, $usersGame->user_id);
     }
 
     /** @test */
@@ -100,19 +94,25 @@ class GameTest extends TestCase
     }
 
     /** @test */
+    public function a_game_can_be_deleted()
+    {
+        $this->signIn();
+        $game = create(Game::class);
+
+        $this->delete($game->path());
+
+        $this->assertDatabaseMissing('games', $game->toArray());
+    }
+
+    /** @test */
     public function when_a_game_is_deleted_all_of_its_associated_frames_should_be_deleted()
     {
         $this->signIn();
-        $game = create(Game::class, [
-            'user_id' => $this->user->id
-        ]);
-        $frames = create(Frame::class, [
-            'game_id' => $game->id
-        ], 10);
+        $game = create(Game::class);
+        $frames = create(Frame::class, ['game_id' => $game->id], 10);
         $game->frames()->saveMany($frames);
 
-        $this->delete($game->path())
-             ->assertRedirect('/games');
+        $this->delete($game->path());
 
         $this->assertDatabaseMissing('frames', ['game_id' => $game->id]);
     }
@@ -121,26 +121,9 @@ class GameTest extends TestCase
     public function it_should_redirect_you_to_your_index_page_when_you_delete_a_game()
     {
         $this->signIn();
-        //dd($this->user->id);
-        $game = create(Game::class, [
-            'user_id' => $this->user->id,
-        ]);
+        $game = create(Game::class);
 
         $this->delete($game->path())
             ->assertRedirect('/games');
-    }
-
-    /** @test */
-    public function it_should_create_one_frame_with_two_rolls()
-    {
-        $this->signIn();
-        $rolls = make(Roll::class, ['pins' => 1], 2);
-        $game = create(Game::class, ['user_id' => $this->user->id]);
-
-        $this->put($game->path(), [
-            'rolls' => $rolls->toArray()
-        ]);
-
-        $this->assertCount(2, $game->rolls);
     }
 }
