+++ tests/Unit/RollTest.php
@@ -23,12 +23,4 @@ class RollTest extends TestCase
 
         $this->assertDatabaseHas('rolls', $roll->toArray());
     }
-
-    /** @test */
-    public function a_roll_should_have_a_pins_key()
-    {
-        $roll = make(Roll::class);
-
-        $this->assertArrayHasKey('pins', $roll->toArray());
-    }
 }
